requirements.txt
Django>=4.0,<5.0
djangorestframework>=3.14,<4.0
mysqlclient>=2.1,<3.0  # For MySQL support.
# psycopg2-binary>=2.9,<3.0  # For PostgreSQL support.